+++ cedar-policy/src/integration_testing.rs
@@ -25,10 +25,11 @@
 #![allow(clippy::unwrap_used)]
 // PANIC SAFETY: This module is used only for testing.
 #![allow(clippy::expect_used)]
+#![allow(deprecated)] // Functionality in this module depends on `frontend`. https://github.com/cedar-policy/cedar/issues/821
 
 use crate::{
-    frontend::is_authorized::InterfaceResponse, AuthorizationError, Authorizer, Context, Decision,
-    Entities, EntityUid, PolicyId, PolicySet, Request, Schema, ValidationMode, Validator,
+    AuthorizationError, Authorizer, Context, Decision, Entities, EntityUid, PolicyId, PolicySet,
+    Request, Schema, ValidationMode, Validator,
 };
 use cedar_policy_core::jsonvalue::JsonValueWithNoDuplicateKeys;
 use serde::{Deserialize, Serialize};
@@ -161,7 +162,7 @@ pub trait CustomCedarImpl {
         q: &cedar_policy_core::ast::Request,
         p: &cedar_policy_core::ast::PolicySet,
         e: &cedar_policy_core::entities::Entities,
-    ) -> InterfaceResponse;
+    ) -> crate::frontend::is_authorized::InterfaceResponse;
 
     /// Custom validator entry point.
     ///