+++ cedar-testing/src/cedar_test_impl.rs
@@ -19,7 +19,10 @@
 //! running the integration tests and for performing randomized differential
 //! testing (see <https://github.com/cedar-policy/cedar-spec>).
 
+#![allow(deprecated)] // Functionality in this module depends on `frontend`. https://github.com/cedar-policy/cedar/issues/821
+
 pub use cedar_policy::frontend::is_authorized::InterfaceResponse;
+
 use cedar_policy_core::ast::{Expr, PolicySet, Request, Value};
 use cedar_policy_core::authorizer::Authorizer;
 use cedar_policy_core::entities::Entities;
@@ -70,7 +73,7 @@ pub struct Micros(pub u128);
 #[derive(Debug, Deserialize)]
 pub struct TestResponse {
     /// Actual response
-    pub response: InterfaceResponse,
+    pub response: cedar_policy::frontend::is_authorized::InterfaceResponse,
     /// Timing info in microseconds. This field is a `HashMap` to allow timing
     /// multiple components (or none at all).
     pub timing_info: HashMap<String, Micros>,
@@ -200,7 +203,8 @@ impl CedarTestImplementation for RustEngine {
         // Error messages should only include the policy id to use the
         // `ErrorComparisonMode::PolicyIds` mode.
         let response = cedar_policy::Response::from(response);
-        let response = InterfaceResponse::new(
+        #[allow(deprecated)]
+        let response = cedar_policy::frontend::is_authorized::InterfaceResponse::new(
             response.decision(),
             response.diagnostics().reason().cloned().collect(),
             response